+++ castor.php
@@ -11,6 +11,7 @@ import(__DIR__ . '/tools/phpstan/castor.php');
 import(__DIR__ . '/tools/static/castor.php');
 
 mount(__DIR__ . '/tools/phar');
+mount(__DIR__ . '/tools/release');
 mount(__DIR__ . '/tools/watcher');
 
 #[AsTask(description: 'hello')]