+++ examples/run.php
@@ -55,13 +55,13 @@ function testFile(): int
 }
 
 #[AsTask(description: 'Run a command that will fail')]
-function exception(): void
+function exception(int $repeat = 1): void
 {
     if (!output()->isVerbose()) {
         output()->writeln('Re-run with -v, -vv, -vvv for different output.');
     }
 
-    run('echo foo; echo bar>&2; exit 1', context: context()->withPty(false)->withQuiet());
+    run('echo $foo; echo ' . str_repeat('bar', $repeat) . '>&2; exit 1', context: context()->withPty(false)->withQuiet()->withEnvironment(['foo' => 'bar']));
 }
 
 #[AsTask(description: 'Run a sub-process and display information about it, with ProcessHelper')]